% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_brms.R
\name{zar_brms}
\alias{zar_brms}
\title{zar_brms}
\usage{
zar_brms(
  name,
  formula,
  prior,
  family = NULL,
  data,
  chains = 4,
  iter = 2000,
  cores = 1,
  init = NULL,
  control = NULL,
  save_model = NULL,
  backend = getOption("brms.backend", "cmdstanr"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  format = "qs",
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  cue = targets::tar_option_get("cue")
)
}
\arguments{
\item{name}{}

\item{formula}{}

\item{prior}{}

\item{data}{}

\item{chains}{}

\item{iter}{}

\item{control}{}

\item{save_model}{}
}
\description{
zar_brms
}
