% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_tracks.R
\name{zar_tracks}
\alias{zar_tracks}
\title{zar_tracks}
\usage{
zar_tracks(
  data,
  crs,
  rate,
  tolerance,
  n_random_steps,
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  format = "qs",
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  cue = targets::tar_option_get("cue")
)
}
\value{

}
\description{
zar_tracks
}
